refactor: Change Product name into static product and single product into dynamic product
The reason of using these name is to identify the product data is being fetched or not. The name static is used if user is available to get single product using existing products data.
The name dynamic is being used if there is no static single product(have to fetch the new one).